Fixed exception when manually increasing size of inspector collection:
When manually increasing the size of a collection in the inspector, serialization calls happen when the provided data still only has the old collection size, resulting in an OutOfRangeException. This change caches the previous (last good) element from the enumeration, and returns it in the event that the enumerator runs out before reaching the requested index.